// apps/web/src/features/auth/auth.types.ts
export interface User {
    id: number;
    email: string;
    name: string;
    createdAt: string;
}

export interface LoginRequest {
    email: string;
    password: string;
}

export interface LoginResponse {
    token: string;
    user: User
}

export interface RegisterRequest {
    email: string;
    password: string;
    name: string;
}

export interface RegisterResponse {
    id: number;
    email: string;
    name: string;
}